Add Zerocmp Utility
Add a utility method for quickly determining whether a potentially large set of integral values is all zero. This will be important for evaluating whether a source is potentially hardware accelerated, as well as whether a source may not currently be rendering itself. This implementation won in performance tests vs memcmp'ing against an all-zero vector, basic loops, and all_of. The runtime for data with non-black pixels reasonably near the beginning was incredibly small (0.001ms), it could take up to 2ms on large, all-black data sets and should not be run every frame against such.