Finish Freeze Handling, Accel Skeleton
Finished handling of frozen windows, providing indication and explanation of what's happening to the user as well as a cached image from before it froze.

Drafted out when hardware acceleration detection takes place & some of how it will be captured in the code, implementation of actual windows graphics capture code is next.